Created config.py in auth package and moved email configuration there
