+++ api/src/auth/schemas.py
@@ -1,24 +1,8 @@
 from typing import List, Optional
 
-from api.config import EMAIL, EMAIL_PASSWORD
-from fastapi_mail import ConnectionConfig
 from fastapi_users import schemas
 from pydantic import BaseModel, EmailStr
 from sqlalchemy import Integer
-
-conf = ConnectionConfig(
-    MAIL_USERNAME=EMAIL,
-    MAIL_PASSWORD=EMAIL_PASSWORD,
-    MAIL_FROM=EMAIL,
-    MAIL_PORT=587,
-    MAIL_SERVER="smtp.gmail.com",
-    MAIL_FROM_NAME="Guide.me",
-    MAIL_STARTTLS=True,
-    MAIL_SSL_TLS=False,
-    USE_CREDENTIALS=True,
-    VALIDATE_CERTS=True,
-    SUPPRESS_SEND=False,
-)
 
 
 class UserCreate(schemas.BaseUserCreate):
